Refactored to improve output formats
- Fixed bug in batching logic
- Refactored to enable improved output format logic
- Added output format logic to tracker to pass to response handling
- Added output format logic in arguments
- Adapted request handlers to deal with output format logic
- Improved output handling in print/file response handlers to use
additional flags
- Translated all batch sizing to consistent u32
- Created Properties wrapper to better represent geojson properties
transfer

+++ gm-proximity/src/server/handle/bench.rs
@@ -19,5 +19,5 @@ pub fn bench(handler: Context, bench: BenchReq) {
         handler.send(Response::Bench(bench_res));
     }
 
-    handler.send(Response::Done(bench.ratio as usize));
+    handler.send(Response::Done(bench.ratio));
 }
